add leaderboard game state
