+++ common/CastAPI.ts
@@ -1,6 +1,5 @@
 import { GameState } from 'app/common/gameState';
 import { GameSettings } from 'app/common/gameSettings';
-import { LeaderboardEntry } from 'stores/leaderboard-store';
 
 export type CastAPI = CastSenderAPI & CastReceiverAPI;
 
@@ -10,7 +9,6 @@ export interface CastSenderAPI {
   updateGameSettings: (settings: GameSettings) => void;
   updateLocale: (locale: string) => void;
   updateControllers: (controllers: Record<string, string>) => void;
-  updateLeaderboard: (entries: LeaderboardEntry[]) => void;
 }
 
 type Callback<F> = (callback: F) => void;
@@ -25,5 +23,4 @@ export interface CastReceiverAPI {
   onGameSettingsUpdate: CastCallback<'updateGameSettings'>;
   onLocaleUpdate: CastCallback<'updateLocale'>;
   onControllerUpdate: CastCallback<'updateControllers'>;
-  onLeaderboardUpdate: CastCallback<'updateLeaderboard'>;
 }
