+++ src-electron/castAPI/preload.ts
@@ -13,7 +13,6 @@ const senderAPI: CastSenderAPI = {
   updateGameSettings: send('updateGameSettings'),
   updateLocale: send('updateLocale'),
   updateControllers: send('updateControllers'),
-  updateLeaderboard: send('updateLeaderboard'),
 };
 
 const on = <K extends keyof CastReceiverAPI>(name: K): CastReceiverAPI[K] => {
@@ -30,7 +29,6 @@ const receiverAPI: CastReceiverAPI = {
   onGameSettingsUpdate: on('onGameSettingsUpdate'),
   onLocaleUpdate: on('onLocaleUpdate'),
   onControllerUpdate: on('onControllerUpdate'),
-  onLeaderboardUpdate: on('onLeaderboardUpdate'),
 };
 
 const api: CastAPI = {
