+++ src/i18n/en-US/index.ts
@@ -10,6 +10,7 @@ export default {
   action: {
     start: 'Start',
     devices: 'Devices',
+    leaderboard: 'Leaderboard',
   },
 
   cast,
