+++ src/stores/cast-store.ts
@@ -4,7 +4,6 @@ import { ref } from 'vue';
 import { GameState } from 'app/common/gameState';
 import { useRouter } from 'vue-router';
 import { GameSettings } from 'app/common/gameSettings';
-import { LeaderboardEntry } from 'stores/leaderboard-store';
 
 export const useCastStore = defineStore('cast', () => {
   const { locale } = useI18n();
@@ -48,22 +47,15 @@ export const useCastStore = defineStore('cast', () => {
     locale.value = value;
   }
 
-  const leaderboard = ref<LeaderboardEntry[]>();
-  function updateLeaderboard(entries: LeaderboardEntry[]) {
-    leaderboard.value = entries;
-  }
-
   return {
     controllers,
     gameState,
     gameSettings,
-    leaderboard,
 
     updateGameState,
     updateGameSettings,
     updateLocale,
     updateControllers,
-    updateLeaderboard,
   };
 });
